Got syntax right (hopefully).
The difficulty was integrating a partial order in which ''term and ''atomic are treated just like the other syntactic categories, with the fact that in the grammar rules for Params, Atomic and Term need to be parsed in lockstep. To achieve that, we break down Atomic into Atomic-base-atomic and Atomic-greater-atomic. Furthermore, we break down Term into Term-base-nonatomic, Term-greater-nonatomic, and Term-greater-atomic. The same breakdown must be done for the other syntactic categories as well.